pollsapi/requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
astroid==2.0.2
autopep8==1.3.5
backcall==0.1.0
base58==0.2.5
bleach==2.1.3
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptoconditions==0.7.0
cryptography==2.3
Cython==0.28.5
decorator==4.3.0
Django==2.1
django-cors-headers==2.4.0
django-extensions==2.1.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
entrypoints==0.2.3
html5lib==1.0.1
idna==2.7
ipykernel==4.8.2
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.3.2
isort==4.3.4
itypes==1.1.0
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.6.0
numpy==1.15.0
openapi-codec==1.3.2
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
pipenv==2018.7.1
prometheus-client==0.3.1
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pyasn1==0.4.4
pycodestyle==2.4.0
pycparser==2.18
Pygments==2.2.0
pylint==2.1.0
pylint-django==2.0
pylint-plugin-utils==0.4
pymongo==3.7.1
PyNaCl==1.1.2
pysha3==1.0.2
python-dateutil==2.7.3
python-rapidjson==0.6.3
pytz==2018.5
pyzmq==17.1.0
qtconsole==4.3.1
requests==2.19.1
rethinkdb==2.3.0.post6
scikit-learn==0.19.2
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
simplejson==3.16.0
six==1.11.0
sklearn==0.0
terminado==0.8.1
testpath==0.3.1
tornado==5.1
traitlets==4.3.2
typed-ast==1.1.0
typing==3.6.4
uritemplate==3.0.0
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.3.1
wrapt==1.10.11
